MLproject_Solar_Irradiance/requirements.txt
pydantic
requests
ttkthemes
tkintermapview
matplotlib
numpy
pandas
gunicorn
dash
pandas
Pillow
